feat: feature-flag PyO3, expose public Rust API (closes #6)
- Make `pyo3` an optional dependency behind the `python` feature flag
  (enabled by default); `cargo check --no-default-features` now succeeds
  without pulling in PyO3.
- Move all PyO3 bindings into `src/bindings/python.rs` (gated with
  `#![cfg(feature = "python")]`) and wire them via `src/bindings/mod.rs`.
- Rewrite `src/lib.rs` as a clean public Rust API: `parse()`,
  `AstNode`, `Token`, `TokenKind`, and `ParseError` are all `pub`.
- Re-export the `#[pymodule] _rust` entry point from `lib.rs` under
  `#[cfg(feature = "python")]` so maturin continues to find it.
- `maturin develop --uv && pytest tests/` — all 1718 Python tests pass.
