+++ Cargo.toml
@@ -7,5 +7,9 @@ edition = "2021"
 name = "_rust"
 crate-type = ["cdylib"]
 
+[features]
+default = ["python"]
+python = ["dep:pyo3"]
+
 [dependencies]
-pyo3 = { version = "0.25", features = ["extension-module"] }
+pyo3 = { version = "0.25", features = ["extension-module"], optional = true }
