+++ src/bindings/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "python")]
+pub mod python;
